fix: add special interface for rate chain
