+++ src/data-access/rateClient/bank.gov.rate.service.ts
@@ -1,9 +1,10 @@
-import {BankGovRateData, IRateService} from "../../service/services/rate/rate.service.interface";
+import {BankGovRateData} from "./rate.client.interface";
 import axios from "axios";
 import logger from "../../helpers/logger";
 import {BANK_GOV_URL, CURRENCY} from "../config/rate.api.const";
+import {IRateChainService} from "../../service/services/rate/client/chain";
 
-export class BankGovRateService implements IRateService{
+export class BankGovRateService implements IRateChainService{
 
     async getExchangeRate(): Promise<number> {
         const response = await axios.get<BankGovRateData[]>(BANK_GOV_URL);
