+++ src/data-access/rateClient/exchangeapi.rate.service.ts
@@ -1,9 +1,10 @@
-import {Fawazahmed0ExchangeRateData, IRateService} from "../../service/services/rate/rate.service.interface";
+import {Fawazahmed0ExchangeRateData} from "./rate.client.interface";
 import axios from "axios";
 import logger from "../../helpers/logger";
 import {EXCHANGE_API_URL} from "../config/rate.api.const";
+import {IRateChainService} from "../../service/services/rate/client/chain";
 
-export class ExchangeAPIRateService implements IRateService{
+export class ExchangeAPIRateService implements IRateChainService{
 
     async getExchangeRate(): Promise<number> {
         const response = await axios.get<Fawazahmed0ExchangeRateData>(EXCHANGE_API_URL);
