+++ src/router/controllers/rate.controller.ts
@@ -1,6 +1,6 @@
 import {Request, Response} from 'express'
 import logger from '../../helpers/logger'
-import {IRateService} from "../../service/services/rate/rate.service.interface";
+import {IRateService} from "../../service/services/rate/rate.service.interface"
 import rateService from "../../service/services/rate/rate.service";
 import {errorHandler} from "../../error/handler/error.handler";
