+++ src/service/services/rate/rate.service.interface.ts
@@ -1,23 +1,3 @@
 export interface IRateService {
     getExchangeRate(): Promise<number>;
 }
-
-export interface PrivateBankRateData {
-    ccy: string;
-    base_ccy: string;
-    buy: number;
-    sale: number;
-}
-
-export interface BankGovRateData {
-    rate: number;
-    cc: string
-}
-
-export interface CurrencyRates {
-    [key: string]: number;
-}
-
-export interface Fawazahmed0ExchangeRateData {
-    usd: CurrencyRates
-}
